fix(gameobjects/flags): correct ObjF* enum backing types to uint
All game object property flag fields are read from a 4-byte uint32 on
the wire. Six flag enums were declared with smaller backing types that
could silently truncate high bits if the engine ever sets them.

- ObjFArmorFlags  : byte  -> uint
- ObjFBlitFlags   : byte  -> uint
- ObjFContainerFlags : ushort -> uint
- ObjFPortalFlags : ushort -> uint
- ObjFSceneryFlags: ushort -> uint
- ObjFWeaponFlags : ushort -> uint

+++ src/GameObjects/ArcNET.GameObjects/Flags/ObjFArmorFlags.cs
@@ -2,7 +2,7 @@
 
 /// <summary>Armor-specific flags (OARF_).</summary>
 [Flags]
-public enum ObjFArmorFlags : byte
+public enum ObjFArmorFlags : uint
 {
     /// <summary>No flags set.</summary>
     None = 0,
